Move metafilter package up
