+++ src/main/java/tudelft/ti2806/pl3/ControllerContainer.java
@@ -1,6 +1,6 @@
 package tudelft.ti2806.pl3;
 
-import tudelft.ti2806.pl3.data.metafilter.MetaFilterController;
+import tudelft.ti2806.pl3.metafilter.MetaFilterController;
 import tudelft.ti2806.pl3.findgenes.FindgenesController;
 import tudelft.ti2806.pl3.sidebar.SideBarController;
 import tudelft.ti2806.pl3.sidebar.phylotree.PhyloController;
